Allow tests to be run locally (without a container)
Add ``use_container`` and ``command_prefix``, so tests can be run on the host,
rather that on a container running on the host.
This allows testing of the services running either in the containers or
running baremetal.

To run tests bare_metal, the ``use_container`` var must be set.
This will set-up the ``command_prefix`` var to be '', instead of
``{{ container_bin }} exec {{ container_name }}``.